Model to table

This function takes a model object and generates a table in pdf or word format.

mod_to_table(
  model,
  param_names = NULL,
  font_size = 11,
  left_align = 1,
  width = NULL,
  col = NULL,
  ci = 0.95,
  rope = c(-0.1, 0.1)
)

Arguments

model
A model object

param_names
For renaming the default parameter names

font_size
The font size for the table

left_align
Columns that should be left aligned

width
The width in inches of the columns

col
The number of a specific column for width adjustments

ci
Range of the HDCI for brms models

rope
Range of the ROPE interval for brms models

Examples

mod <- lm(mpg ~ drat, data = mtcars)
mod_to_table(mod)

#> \begin{table}[!h]
#> \centering\begingroup\fontsize{11}{13}\selectfont
#> 
#> \begin{tabular}{@{}lrrrrrr@{}}
#> \toprule
#> Parameter & Estimate & SE & CI low & CI high & \emph{t} & \emph{p}\\
#> \midrule
#> (Intercept) & -7.52 & 5.48 & -18.71 & 3.66 & -1.37 & .18\\
#> drat & 7.68 & 1.51 & 4.60 & 10.76 & 5.10 & < .001\\
#> \bottomrule
#> \end{tabular}
#> \endgroup{}
#> \end{table}